requirements.txt
setuptools~=68.2.0
python-telegram-bot==21.1.1
mypy